+++ app/src/main/java/com/softbankrobotics/chatbotsample/ChatbotUtteredAndAnimatedReaction.java
@@ -11,10 +11,10 @@ import com.aldebaran.qi.Future;
 import com.aldebaran.qi.sdk.QiContext;
 import com.aldebaran.qi.sdk.builder.AnimateBuilder;
 import com.aldebaran.qi.sdk.builder.AnimationBuilder;
+import com.aldebaran.qi.sdk.builder.SayBuilder;
 import com.aldebaran.qi.sdk.object.actuation.Animate;
 import com.aldebaran.qi.sdk.object.actuation.Animation;
 import com.aldebaran.qi.sdk.object.conversation.BaseChatbotReaction;
-import com.aldebaran.qi.sdk.object.conversation.Phrase;
 import com.aldebaran.qi.sdk.object.conversation.Say;
 import com.aldebaran.qi.sdk.object.conversation.SpeechEngine;
 
@@ -43,9 +43,12 @@ public class ChatbotUtteredAndAnimatedReaction extends BaseChatbotReaction {
     @Override
     public void runWith(final SpeechEngine speechEngine) {
 
-        // All Say actions that must be executed during the call of the this method must be created
-        // via the SpeechEngine.
-        Say say = speechEngine.makeSay(new Phrase(toBeSaid));
+        // All Say actions that must be executed inside this method must be created
+        // via the provided SpeechEngine.
+
+        Say say = SayBuilder.with(speechEngine)
+                            .withText(toBeSaid)
+                            .build();
 
         Animation animation = AnimationBuilder.with(getQiContext())
                                               .withResources(animationResourceId)
@@ -54,14 +57,14 @@ public class ChatbotUtteredAndAnimatedReaction extends BaseChatbotReaction {
                                         .withAnimation(animation)
                                         .build();
 
-        // The actions must be executed asynchronously in order to get futures that can be
-        // canceled by the head thanks to the stop() method.
-        // Additionally, here it allows both actions to be executed in parallel.
+        // The actions must be executed asynchronously in order to get futures that allow
+        // the actions to be canceled by the chat engine thanks to the stop() method.
+        // Additionally, the asynchronous call allows here both actions to be executed in parallel.
         fsay = say.async().run();
         fanimate = animate.async().run();
 
         try {
-            // One must not leave runWith before the actions are terminated : thus wait on the futures
+            // One must not leave runWith before the actions are terminated: thus wait on the futures
             fsay.get();
             fanimate.get();
         } catch (ExecutionException e) {