+++ CHANGELOG.md
@@ -6,7 +6,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
 [Unreleased]
 
-## [0.1] - 2018-05-30
+## [0.0.2] - 2018-06-06
+
+### changed
+- Now using TopicBuilder
+- DialogflowChatbot no longer needs Robot
+- renamed 'Class diagram.plantuml' into 'classDiagram.plantuml'
+- Improved README.md
+- upgrade to QiSDK 1.3.3
+- MainActivity inherits RobotActivity
+- Renaming the action string into 'excitement'
+- Dialogflow agent no longer only in charge of the small talk, explain it in README
+- Providing the Dialogflow agent as a zip file that can be imported
+- Now using SayBuilder.withSpeechEngine()
+- Bump versionCode to 2
+
+## [0.0.1] - 2018-05-30
 
 ### added
 - Initial version of the project